requirements.txt
beautifulsoup4
bs4
certifi
charset-normalizer
Flask
gevent
greenlet
gunicorn
idna
jieba
Jinja2
kiwisolver
MarkupSafe
matplotlib
numpy
Pillow
pyparsing
python-dateutil
requests
six
soupsieve
urllib3
Werkzeug
wordcloud
